Register nodes with monzo.com/cpu-period resource
We have a custom pod resource which allows tuning the CPU throttling
period. Upgrading to 1.9 causes this to break scheduling logic, as the
scheduler and pod preemption controller takes this resource into account
when deciding where to place pods, and which pods to pre-empt.

This patches the kubelet so that it registers its node with a large
amount of this resource - 10000 * max number of pods (default 110). We
typically run pods with this set to 5000, so this should be plenty.